[MacOS] Basic ListViewBaseInternalContainer implementation
No support for dynamic item sizes is provided yet.
