+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/ListViewBaseSource.iOSMacOS.cs
@@ -8,6 +8,10 @@ using Uno.Extensions;
 using Windows.Foundation;
 using Windows.UI.Xaml.Data;
 using Uno.Logging;
+using System.Globalization;
+using Uno.UI;
+using Windows.UI.Xaml.Controls.Primitives;
+using Uno.Diagnostics.Eventing;
 #if __IOS__
 using UIKit;
 #else
@@ -18,6 +22,17 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class ListViewBaseSource
 	{
+
+		private readonly static IEventProvider _trace = Tracing.Get(TraceProvider.Id);
+
+		public static class TraceProvider
+		{
+			public readonly static Guid Id = Guid.Parse("{EE64E62E-67BD-496A-A8B1-4A142642B3A3}");
+
+			public const int ListViewBaseSource_GetCellStart = 1;
+			public const int ListViewBaseSource_GetCellStop = 2;
+		}
+
 		/// <summary>
 		/// Key used to represent a null DataTemplate in _templateCache and _templateCells dictionaries (because null is a not a valid key) 
 		/// </summary>
@@ -109,6 +124,94 @@ namespace Windows.UI.Xaml.Controls
 				return 0;
 			}
 			return Owner.XamlParent.GetDisplayGroupCount((int)section);
+		}
+#if __IOS__
+		public override UICollectionViewCell GetCell(UICollectionView collectionView, NSIndexPath indexPath)
+#else
+		public override NSCollectionViewItem GetItem(NSCollectionView collectionView, NSIndexPath indexPath)
+#endif
+		{
+			using (
+			   _trace.WriteEventActivity(
+				   TraceProvider.ListViewBaseSource_GetCellStart,
+				   TraceProvider.ListViewBaseSource_GetCellStop
+			   )
+			)
+			{
+				// Marks this item to be materialized, so its actual measured size 
+				// is used during the calculation of the layout.
+				// This is required for paged lists, so that the layout calculation
+				// does not eagerly get all the items of the ItemsSource.
+				UpdateLastMaterializedItem(indexPath);
+
+				var index = Owner?.XamlParent?.GetIndexFromIndexPath(IndexPath.FromNSIndexPath(indexPath)) ?? -1;
+
+				var identifier = GetReusableCellIdentifier(indexPath);
+
+				var listView = (NativeListViewBase)collectionView;
+#if __IOS__
+				var cell = (ListViewBaseInternalContainer)collectionView.DequeueReusableCell(identifier, indexPath);
+#else
+				var cell = (ListViewBaseInternalContainer)collectionView.MakeItem(GetReusableCellIdentifier(indexPath), indexPath);
+#endif
+
+				using (cell.InterceptSetNeedsLayout())
+				{
+					var selectorItem = cell.Content as SelectorItem;
+
+					if (selectorItem == null ||
+						// If it's not a generated container then it must be an item that returned true for IsItemItsOwnContainerOverride (eg an
+						// explicitly-defined ListViewItem), and shouldn't be recycled for a different item.
+						!selectorItem.IsGeneratedContainer)
+					{
+						cell.Owner = Owner;
+						selectorItem = Owner?.XamlParent?.GetContainerForIndex(index) as SelectorItem;
+						cell.Content = selectorItem;
+						if (this.Log().IsEnabled(Microsoft.Extensions.Logging.LogLevel.Debug))
+						{
+							this.Log().Debug($"Creating new view at indexPath={indexPath}.");
+						}
+
+						FrameworkElement.InitializePhaseBinding(selectorItem);
+
+						// Ensure the item has a parent, since it's added to the native collection view
+						// which does not automatically sets the parent DependencyObject.
+						selectorItem.SetParent(Owner?.XamlParent);
+					}
+					else if (this.Log().IsEnabled(Microsoft.Extensions.Logging.LogLevel.Debug))
+					{
+						this.Log().Debug($"Reusing view at indexPath={indexPath}, previously bound to {selectorItem.DataContext}.");
+					}
+
+					Owner?.XamlParent?.PrepareContainerForIndex(selectorItem, index);
+
+					// Normally this happens when the SelectorItem.Content is set, but there's an edge case where after a refresh, a
+					// container can be dequeued which happens to have had exactly the same DataContext as the new item.
+					cell.ClearMeasuredSize();
+				}
+
+				Owner?.XamlParent?.TryLoadMoreItems(index);
+
+				return cell;
+			}
+		}
+
+		/// <summary>
+		/// Update record of the furthest item materialized.
+		/// </summary>
+		/// <param name="newItem">Item currently being materialized.</param>
+		/// <returns>True if the value has changed and the layout would change.</returns>
+		internal bool UpdateLastMaterializedItem(NSIndexPath newItem)
+		{
+			if (newItem.Compare(_lastMaterializedItem) > 0)
+			{
+				_lastMaterializedItem = newItem;
+				return Owner.ItemTemplateSelector != null; ;
+			}
+			else
+			{
+				return false;
+			}
 		}
 
 #if __IOS__ //TODO
@@ -376,6 +479,32 @@ namespace Windows.UI.Xaml.Controls
 		internal void ReloadData()
 		{
 			_lastMaterializedItem = NSIndexPath.FromItemSection(0, 0);
+		}
+
+		/// <summary>
+		/// Determine the identifier to use for the dequeuing.
+		/// This avoid for already materialized items with a specific template
+		/// to switch to a different template, and have to re-create the content 
+		/// template.
+		/// </summary>
+		private NSString GetReusableCellIdentifier(NSIndexPath indexPath)
+		{
+			NSString identifier;
+			var template = GetTemplateForItem(indexPath);
+
+			if (!_templateCells.TryGetValue(template ?? _nullDataTemplateKey, out identifier))
+			{
+				identifier = new NSString(_templateCache.Count.ToString(CultureInfo.InvariantCulture));
+				_templateCells[template ?? _nullDataTemplateKey] = identifier;
+
+#if __IOS__
+				Owner.RegisterClassForCell(typeof(ListViewBaseInternalContainer), identifier);
+#else
+				Owner.RegisterClassForItem(typeof(ListViewBaseInternalContainer), identifier);
+#endif
+			}
+
+			return identifier;
 		}
 
 		/// <summary>
